Finalize "mixin"-like structure over monolithic included code
