+++ gengo/Cargo.toml
@@ -34,11 +34,9 @@ serde = { version = "1", features = ["derive"] }
 
 [build-dependencies]
 indexmap = { version = "2", features = ["serde"] }
-prettyplease = "0.2"
 proc-macro2 = "1"
 serde_json.workspace = true
 serde_yaml = "0.9"
-syn = "2"
 quote = "1"
 
 [dev-dependencies]
